separate module for the filelogger
